resample nodes if not same as batch_size

+++ moge/network/heterogeneous_network.py
@@ -230,7 +230,7 @@ class HeterogeneousNetwork(NetworkTrainTestSplit):
                          d['type'] in edge_types]
             adj = nx.normalized_laplacian_matrix(nx.Graph(incoming_graph_data=edge_list), nodelist=node_list)
 
-        return adj.astype(float)
+        return adj
 
     def sample_random_negative_edges(self, pos_adj, negative_sampling_ratio):
         pos_rows, pos_cols = pos_adj.nonzero()
